json: use int instead of uint64_t for array indexes
Unsigned types are like a plague - should you use it once, and it will
quickly spread throughout your whole code base, because comparing an
unsigned value with a signed one without type conversion will make the
compiler whine. However, they are less efficient, because the compiler
has to guarantee that integer overflow works predictably for them.
That said let's make json_token::num signed. Let's also use a plain int
for it rather than int64_t, because it's highly unlikely that the
capacity of int won't be enough to store a tuple index.

+++ src/lib/json/json.c
@@ -136,7 +136,7 @@ json_parse_integer(struct json_lexer *lexer, struct json_token *token)
 	const char *pos = lexer->src + lexer->offset;
 	assert(pos < end);
 	int len = 0;
-	uint64_t value = 0;
+	int value = 0;
 	char c = *pos;
 	if (! isdigit(c))
 		return lexer->symbol_count + 1;
